requirements.txt
pandas
scikit-learn
numpy
matplotlib
tqdm
requests
python-dotenv